backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi==0.109.1",
    "numpy==1.26.0",
    "pillow==10.1.0",
    "torch==2.2.0",
    "torchvision==0.17.0",
    "ultralytics==8.0.156",
    "uvicorn[standard]==0.23.2",
]
